tangle - extract common window (#141154)
