+++ src/vs/platform/sharedProcess/electron-main/sharedProcess.ts
@@ -19,7 +19,7 @@ import { IProtocolMainService } from 'vs/platform/protocol/electron-main/protoco
 import { ISharedProcess, ISharedProcessConfiguration } from 'vs/platform/sharedProcess/node/sharedProcess';
 import { ISharedProcessWorkerConfiguration } from 'vs/platform/sharedProcess/common/sharedProcessWorkerService';
 import { IThemeMainService } from 'vs/platform/theme/electron-main/themeMainService';
-import { WindowError } from 'vs/platform/windows/electron-main/windows';
+import { WindowError } from 'vs/platform/window/electron-main/window';
 import { toErrorMessage } from 'vs/base/common/errorMessage';
 
 export class SharedProcess extends Disposable implements ISharedProcess {
