+++ src/vs/platform/window/common/window.ts
