+++ src/vs/platform/windows/test/electron-main/windowsFinder.test.ts
@@ -13,8 +13,8 @@ import { URI } from 'vs/base/common/uri';
 import { getPathFromAmdModule } from 'vs/base/test/node/testUtils';
 import { ICommandAction } from 'vs/platform/actions/common/actions';
 import { NativeParsedArgs } from 'vs/platform/environment/common/argv';
-import { INativeWindowConfiguration } from 'vs/platform/windows/common/windows';
-import { ICodeWindow, ILoadEvent, IWindowState } from 'vs/platform/windows/electron-main/windows';
+import { INativeWindowConfiguration } from 'vs/platform/window/common/window';
+import { ICodeWindow, ILoadEvent, IWindowState } from 'vs/platform/window/electron-main/window';
 import { findWindowOnFile } from 'vs/platform/windows/electron-main/windowsFinder';
 import { toWorkspaceFolders } from 'vs/platform/workspaces/common/workspaces';
 import { IWorkspaceIdentifier } from 'vs/platform/workspace/common/workspace';
