+++ src/vs/platform/windows/test/electron-main/windowsStateHandler.test.ts
@@ -7,7 +7,7 @@ import * as assert from 'assert';
 import { tmpdir } from 'os';
 import { join } from 'vs/base/common/path';
 import { URI } from 'vs/base/common/uri';
-import { IWindowState as IWindowUIState, WindowMode } from 'vs/platform/windows/electron-main/windows';
+import { IWindowState as IWindowUIState, WindowMode } from 'vs/platform/window/electron-main/window';
 import { getWindowsStateStoreData, IWindowsState, IWindowState, restoreWindowsState } from 'vs/platform/windows/electron-main/windowsStateHandler';
 import { IWorkspaceIdentifier } from 'vs/platform/workspace/common/workspace';
