+++ src/vs/platform/workspaces/electron-main/workspacesHistoryMainService.ts
@@ -21,7 +21,7 @@ import { ILogService } from 'vs/platform/log/common/log';
 import { IStateMainService } from 'vs/platform/state/electron-main/state';
 import { StorageScope, StorageTarget } from 'vs/platform/storage/common/storage';
 import { IGlobalStorageMainService } from 'vs/platform/storage/electron-main/storageMainService';
-import { ICodeWindow } from 'vs/platform/windows/electron-main/windows';
+import { ICodeWindow } from 'vs/platform/window/electron-main/window';
 import { IRecent, IRecentFile, IRecentFolder, IRecentlyOpened, IRecentWorkspace, isRecentFile, isRecentFolder, isRecentWorkspace, RecentlyOpenedStorageData, restoreRecentlyOpened, toStoreData } from 'vs/platform/workspaces/common/workspaces';
 import { isSingleFolderWorkspaceIdentifier, isWorkspaceIdentifier, IWorkspaceIdentifier, WORKSPACE_EXTENSION } from 'vs/platform/workspace/common/workspace';
 import { IWorkspacesManagementMainService } from 'vs/platform/workspaces/electron-main/workspacesManagementMainService';
